pycryptodome/pyproject.toml
[project]

name = "pycryptodome_app"
version = "1.0.0"
description = ""
authors = [{name = "Flet Developer", email = "your@email.com"}]
dependencies = ["flet==0.25.1", "pycryptodome"]

[tool.flet]
android.adaptive_icon_background = "#ffffff"